// frontend/src/types/OVMSTypes.ts
type OVMSData = {
    id: number;
    timestamp: Date;
    operative_state: string;
    latitude: number;
    longitude: number;
    elevation: number;
    slope: number;
    speed: number;
    odometer: number;
    batt_temp: number;
    ext_temp: number;
    power_kw: number;
    capacity: number;
    vehicle_id: string;
    soc: number;
    soh: number;
    voltage: number;
    current: number;
    throttle: number;
    regen_brake: number;
    consumption: number;
    range_est: number;
    range_ideal: number;
    range_full: number;
    drivetime: number;
    charge_time: number;
    footbrake: boolean;
    engine_temp: number;
    is_charging: boolean;
    tpms: string;
    coulomb: number;
    energy: number;
    rpm: number;
    charger_type: string;
    coulomb_rec: number;
    drivemode: string;
    energy_rec: number;
    mass: number;
    mec_power: number;
    mean_acc: number;
    friction_force: number;
    net_force: number;
    run: number;
    mean_speed: number;
    freeram: number;
    net_signal: number;
    tasks: number;
    elevation2: number;
    charge_current: number;
    angle_x: number;
    angle_y: number;
    AcX: number;
    AcY: number;
    AcZ: number;
    humidity: number;
    kwh_km: number;
    assist_level: number;
    pressure: number;
    user_name: string;
    user_id: number;
    acceleration: number;
}

export type { OVMSData };